(ranran) FIX: restore all open convoi and halt windows
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@10759 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/tool/simtool.cc
@@ -7067,8 +7067,15 @@ bool tool_rotate90_t::init( player_t * )
 
 bool tool_quit_t::init( player_t * )
 {
-	destroy_all_win( true );
-	welt->stop( strempty(default_param) );
+	if (!strempty(default_param)) {
+		// new world
+		destroy_all_win(true);
+		welt->stop(false);
+	}
+	else {
+		// totally quit
+		welt->stop(true);
+	}
 	return false;
 }
